pyproject.toml
[project]
name = "strands-chainlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.0",
    "strands-agents>=0.2.1",
    "strands-agents-tools>=0.1.8",
]
